Remove circular import
https://github.com/fedify-dev/drfed/pull/47#discussion_r3869714565
